pyproject.toml
[project]
name = "edvinasetherel"
version = "1.0.0"
requires-python = "==3.14"
dependencies = [
    "pydantic==2.12.5",
    "pyyaml==6.0.3",
]
